+++ flask_restly/decorator/put.py
@@ -1,9 +1,8 @@
 from flask_restly._storage import append_mapping
 from functools import wraps
-from flask import jsonify
 
 
-def put(path):
+def put(path, serializer=None):
     def decorator(func):
         @wraps(func)
         def wrapper(*args, **kwargs):
@@ -13,13 +12,13 @@ def put(path):
             assert length <= 2, 'Too much return items in PUT method'
 
             if length <= 1:
-                return jsonify(response), 201
+                return response, 201
 
             assert isinstance(response[1], int) is True, 'PUT response code should be a number'
 
-            return jsonify(response[0]), response[1]
+            return response[0], response[1]
 
-        append_mapping(wrapper, path, 'PUT')
+        append_mapping(wrapper, path, serializer, 'PUT')
 
         return wrapper