+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.0 (UNRELEASED)
+* Ability to skip `flask-restly`'s serialization when `Response` object is used
+* Custom serializer support
+
 ## 0.1.1 (12.10.2018)
 * Fixed PyPi markdown README view